fixed_requirements.txt
asttokens==2.4.1
attrs==23.2.0
autograd==1.6.2
autograd-gamma==0.5.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
ConfigSpace @ git+https://github.com/CrunchyFlakes/ConfigSpace@127a106141824a7293d4cc4ddcd463f47d03a9b0
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
dask==2024.7.1
dask-jobqueue==0.8.5
decorator==5.1.1
distributed==2024.7.1
emcee==3.1.6
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
fire==0.6.0
focal-loss-torch==0.1.2
fonttools==4.53.1
formulaic==1.0.2
fsspec==2024.6.1
future==1.0.0
fvcore==0.1.5.post20221221
gdown==5.2.0
GraKeL==0.1.10
graphviz==0.20.3
idna==3.7
imageio==2.34.2
importlib_metadata==8.2.0
importlib_resources==6.4.0
iniconfig==2.0.0
interface-meta==1.3.0
iopath==0.1.10
ipython==8.18.1
ipywidgets==8.1.3
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
lazy_loader==0.4
lifelines==0.29.0
lightgbm==4.5.0
locket==1.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
medmnist==3.0.1
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
nasbench-pytorch==1.3.1
-e git+https://github.com/CrunchyFlakes/NASLib@7b93550e09e766418ed350c5066825c43b06c82f#egg=naslib&subdirectory=../../NASLib
nbformat==5.10.4
networkx==3.2.1
ngboost==0.5.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
opt-einsum==3.3.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
plotly==5.23.0
pluggy==1.5.0
portalocker==2.10.1
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pybnn==0.0.5
pydotz==1.5.1
Pygments==2.18.0
pynisher==1.0.10
pyparsing==3.1.2
pyrfr==0.9.0
pyro-api==0.1.2
pyro-ppl==1.9.1
PySocks==1.7.1
pytest==8.3.2
python-dateutil==2.9.0.post0
pytorch-msssim==1.0.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rpds-py==0.19.1
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
smac==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
tblib==3.0.0
tenacity==9.0.0
tensorwatch==0.9.1
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.7.24
tomli==2.0.1
toolz==0.12.1
torch==2.4.0
torchinfo==1.8.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transforms3d==0.4.2
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
widgetsnbextension==4.0.11
wrapt==1.16.0
xgboost==2.1.1
yacs==0.1.8
zict==3.0.0
zipp==3.19.2
